cloud_functions/pyproject.toml
[project]
name = "grantflow-cloud-functions"
version = "0.1.0"
description = "Cloud functions for GrantFlow.AI"
authors = [
  { name = "GrantFlow.AI Team", email = "team@grantflow.ai" },
]
requires-python = ">=3.13"
classifiers = [ "Programming Language :: Python :: 3 :: Only", "Programming Language :: Python :: 3.13" ]
dependencies = [
  "asyncpg>=0.29",
  "cloudevents>=1.10",
  "firebase-admin>=6.5",
  "firebase-functions>=0.1.1",
  "functions-framework>=3.5",
  "google-cloud-firestore>=2.11",
  "google-cloud-pubsub>=2.18",
  "httpx>=0.27",
  "jinja2>=3.1",
  "markdown>=3.5",
  "python-docx>=1.1",
  "sqlalchemy>=2",
]

[tool.mypy]
python_version = "3.13"
strict = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
  "google.cloud.*",
  "google.cloud.firestore",
  "google.cloud.firestore_v1",
  "google.cloud.pubsub_v1",
  "firebase_admin.*",
  "firebase_functions.*",
  "cloudevents.*",
  "docx.*",
  "markdown",
]
ignore_missing_imports = true
follow_imports = "skip"
no_implicit_reexport = false

[tool.uv]
dev-dependencies = [
  "types-requests>=2.32.0", # Specific to cloud_functions, not in root
]
